implemented InMemoryUserRepository extended Meal from AbstractBaseEntity and added usedId moved MealTo to package 'to' added the check by usedId to InMemoryMealRepository

+++ src/main/java/ru/javawebinar/topjava/repository/inmemory/InMemoryMealRepository.java
@@ -3,11 +3,14 @@ package ru.javawebinar.topjava.repository.inmemory;
 import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
 import ru.javawebinar.topjava.util.MealsUtil;
+import ru.javawebinar.topjava.web.SecurityUtil;
 
 import java.util.Collection;
+import java.util.Comparator;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.stream.Collectors;
 
 public class InMemoryMealRepository implements MealRepository {
     private Map<Integer, Meal> repository = new ConcurrentHashMap<>();
@@ -21,26 +24,40 @@ public class InMemoryMealRepository implements MealRepository {
     public Meal save(Meal meal) {
         if (meal.isNew()) {
             meal.setId(counter.incrementAndGet());
+            meal.setUserId(SecurityUtil.authUserId());
             repository.put(meal.getId(), meal);
             return meal;
         }
         // handle case: update, but not present in storage
-        return repository.computeIfPresent(meal.getId(), (id, oldMeal) -> meal);
+        return isCheckUserId(meal) ?
+                repository.computeIfPresent(meal.getId(), (id, oldMeal) -> meal) :
+                null;
     }
 
     @Override
     public boolean delete(int id) {
-        return repository.remove(id) != null;
+        if (repository.get(id) != null && isCheckUserId(repository.get(id)))
+            return repository.remove(id) != null;
+        else return false;
     }
 
     @Override
     public Meal get(int id) {
-        return repository.get(id);
+        Meal meal = repository.get(id);
+        return isCheckUserId(meal) ? meal : null;
+    }
+
+    public boolean isCheckUserId(Meal meal) {
+        return meal.getUserId() == SecurityUtil.authUserId();
     }
 
     @Override
     public Collection<Meal> getAll() {
-        return repository.values();
+        return repository.values()
+                .stream()
+                .filter(meal -> isCheckUserId(meal))
+                .sorted(Comparator.comparing(Meal::getDateTime).reversed())
+                .collect(Collectors.toList());
     }
 }
